fix(agentic-cli): use pairing-cancelled payload for QR login cancel
Align with the existing CLI PAIRING_CANCELLED_TYPE from the browser cancel path so mm login qr aborts immediately on Mobile Cancel.
